api/pyproject.toml
[project]
name = "llm-adventure-api"
version = "0.1.0"
description = "LLM Adventure API - World-building game backend"
requires-python = ">=3.11"
dependencies = [
    "quart==0.19.4",
    "python-dotenv==1.0.0",
    "aiofiles==23.2.1",
    "hypercorn==0.16.0",
    "langchain==0.3.7",
    "langchain-openai==0.2.8",
    "langchain-core==0.3.21",
    "langchain-postgres==0.0.12",
    "psycopg[binary]==3.2.12",
    "pydantic==2.9.2",
    "alembic==1.13.1",
    "SQLAlchemy>=2.0.36",
    "geoalchemy2==0.14.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
